api/requirements.txt
fastapi
uvicorn[standard]
python-multipart
ultralytics
opencv-python
pillow
